+++ hfut_stu_lib/exception.py
@@ -1,0 +1,9 @@
+from __future__ import unicode_literals
+
+
+class SystemLoginFailed(Exception):
+    pass
+
+
+class IPBanned(SystemLoginFailed):
+    pass